requirements.txt
fastapi
uvicorn[standard]

numpy<2
pandas
scikit-learn
wfdb

pydantic
python-dotenv
requests
psycopg2-binary

matplotlib
pytest